refactor(core): enforce focused module boundaries
