+++ packages/rust/biors-core/src/tokenizer/types.rs
@@ -1,0 +1,34 @@
+use crate::sequence::ResidueIssue;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct TokenizedProtein {
+    pub id: String,
+    pub length: usize,
+    pub alphabet: String,
+    pub valid: bool,
+    pub tokens: Vec<u8>,
+    pub warnings: Vec<ResidueIssue>,
+    pub errors: Vec<ResidueIssue>,
+}
+
+#[derive(Debug, Default, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct ProteinBatchSummary {
+    pub records: usize,
+    pub total_length: usize,
+    pub valid_records: usize,
+    pub warning_count: usize,
+    pub error_count: usize,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct TokenizedFastaInput {
+    pub input_hash: String,
+    pub records: Vec<TokenizedProtein>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct SummarizedFastaInput {
+    pub input_hash: String,
+    pub summary: ProteinBatchSummary,
+}
