+++ scripts/check.sh
@@ -21,8 +21,12 @@ python3 -m py_compile \
   scripts/benchmark_fasta_vs_biopython.py \
   scripts/compare-benchmark-artifacts.py \
   scripts/check-benchmark-artifact.py \
+  scripts/check-module-size.py \
   scripts/check-release-workflow.py \
   scripts/render_benchmark_report.py
+
+echo "==> module size"
+python3 scripts/check-module-size.py
 
 echo "==> benchmark docs"
 scripts/check-benchmark-docs.sh
